requirements.txt
# Core dependencies
pyyaml>=6.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
jsonschema>=4.17.0
tqdm>=4.65.0

# OpenAI
openai>=1.0.0

# DWSIM integration
pythonnet>=3.0.1
pywin32>=300.0 ; sys_platform == 'win32'

# For Linux-specific DWSIM integration
clr-loader>=0.2.0 ; sys_platform == 'linux'
pythonnet-mono>=0.4.0 ; sys_platform == 'linux'

# Utilities
pydantic>=2.0.0
loguru>=0.7.0